Add simplified version of `.json()` method for HttpResponse
CouchDB always respond with UTF-8 encoded JSON, so we don't need in
help of chardet here. All the other cases are extremely rare and in
most of the user will know exact the encoding it have to work with.

But the most import motivation for this fix is that chardet.detect
cannot bytearray and actually treats it as unicode object.