Introduce HTTPSession for sharing auth provider and fork aiohttp.request
Resource instances now may share the same HTTP session, especially if
they are constructed from the same root. The HTTP session is not a very
correct name for a thing, that only holds authentication provider and
TCP connector, but I gave up to figure something better.

The reason of forking aiohttp.request is to simplify the underlying
logic since we don't need to have two places to define default headers
and hold the cookies for aiocouchdb.