Record checkpoints
Checkpoints are special documents that contains all the required
information to let recover replication from in the middle of
the process or continue it in incremental fashion.

The algorithm of doing checkpoints is a bit tricky as we must do
checkpoints incrementally: we cannot record last_seq 300 while 100 is
still in progress.

If checkpoint cannot be made, it's a fatal situation and replication
should be aborted.

couch_replicator also monitors Source and Target peers for
instance_start_time change. This feature will be implemented later.