Remove knowledge about how changes feeds are works
Previous implementation used own iteration over the changes feed in
order to fetch changes. But it makes problematic to restore the feed
in case of connection failure since exception will be raised on
feed.next() yield from. And feed.next() certainly has no idea how to
restart the feed - it's not his job.

So we move that iteration to Peer abstract interface and ask it just
to put the changes into our queue. How it will keep changes feed alive
and works with it is not a replicator business since all it need is
the data from the feed.