Relay on line-based protocol instead of chunk-based one for feeds
CouchDB has a nice hidden feature to stream one logical object per chunk
for views and changes feeds. That was so nice and easy to handle, but...

Suddenly, such strategy tends to be flaky: not all CouchDB-friendly
implementations follows the same rule (CouchBase Sync Gateway), not
all proxies supports the same rule, and even CouchDB 2.0 stop to use
that practice in order to optimize TCP packets fragmentation.

That's how fairy tale came to the end, so we have to switch to readline
protocol. This forced to apply few dirty hacks to make things work.
Suddenly.

This fixes #8 and #9. At least it should.